NGEN-57 Rename algorithm part to algorithm segment
