+++ src/app/generation/core/algorithm-segments/index.ts
