+++ src/app/generation/core/algorithm-segments/template-based-regular-structure-generator.ts
@@ -1,7 +1,7 @@
 import { RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
 import { RegularString } from "@ngen-shared/models";
 import { RandomUtils } from "@ngen-shared/utils";
-import { AlgorithmDataType, AlgorithmPart, GenerationData } from "../models";
+import { AlgorithmDataType, AlgorithmSegment, GenerationData } from "../models";
 
 export interface TemplateBasedRegularStructureGeneratorConfig {
    minLength: number;
@@ -11,7 +11,7 @@ export interface TemplateBasedRegularStructureGeneratorConfig {
    regularNameBase: RegularString;
 }
 
-export class TemplateBasedRegularStructureGenerator extends AlgorithmPart<
+export class TemplateBasedRegularStructureGenerator extends AlgorithmSegment<
    AlgorithmDataType.VOID,
    AlgorithmDataType.REGULAR_STRING_WITH_WILDCARDS,
    TemplateBasedRegularStructureGeneratorConfig
