+++ src/app/generation/core/models/index.ts
@@ -1,6 +1,6 @@
 //@index("./*", (f, _) => `export { ${_.pascalCase(f.name.split(".")[0])} } from "${f.path}";`)
 export { AlgorithmDataType } from "./algorithm-data-type";
-export { AlgorithmPart } from "./algorithm-part";
+export { AlgorithmSegment } from "./algorithm-segment";
 export { GenerationData } from "./generation-data.d";
 export { NameGeneratorAlgorithm } from "./name-generator-algorithm";
 //@endindex
